refa: Nexusのリファクタ
1. 機能分割:
  - 巨大なメインクラスを database, ui, search, list_navigation の4つの責務に分割。
2. ファイル構成:
  - mixins/ フォルダによる整理されたディレクトリ構造。
